chore(tests): update tests
